code-actor/pyproject.toml
[project]
name = "code-actor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "python-dotenv>=1.2.1",
]
